Finish First Main Troll Section
Part 1 of 5 Done

+++ include/party_menu.h
@@ -98,7 +98,8 @@ void GetNumMovesSelectedMonHas(void);
 void MoveDeleterChooseMoveToForget(void);
 // Start qol_field_moves
 // These are all moved from src/party_menu.c
-u8 CanMonLearnTMTutor(struct Pokemon *, u16, u8);
+u8 CanMonLearnTMTutor(struct Pokemon *, u16, u16);
+
 
 enum {
     CAN_LEARN_MOVE,
